Final 2.0

+++ animationApp/View Controllers/MainViewController.swift
@@ -29,8 +29,8 @@ class MainViewController: UIViewController {
     private func setupAnimation(with animationPresets: AnimationModel) {
         animationView.animation = animationPresets.preset
         animationView.curve = animationPresets.curve
-        animationView.force = CGFloat(animationPresets.force)
-        animationView.duration = CGFloat(animationPresets.duration)
+        animationView.force = animationPresets.force
+        animationView.duration = animationPresets.duration
         animationView.delay = animationPresets.delay
         animationView.animate()
     }
